requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.4
httpx==0.26.0